[Tian] 1. Add scene transition for companion scene. 2. Add footprint to pond scene.

+++ VoiceGarden_prototype/VoiceGarden_prototype/Scenes/sceneLevelTwo/Scene_7A.h
@@ -23,7 +23,7 @@
     CCSprite *background;
     CCSprite *door;
     
-    int transitionTime;
+    float transitionTime;
     
 @public
     int sceneStatus;
