+++ VoiceGarden_prototype/VoiceGarden_prototype/Scenes/sceneLevelTwo/Scene_8A.h
@@ -19,7 +19,7 @@
     CCSprite *background;
     CCSprite *door;
     
-    int transitionTime;
+    float transitionTime;
     
 @public
     int sceneStatus;
